+++ Configuration/TCA/tx_oelib_test.php
@@ -6,7 +6,6 @@ return [
     'ctrl' => [
         'tstamp' => 'tstamp',
         'crdate' => 'crdate',
-        'cruser_id' => 'cruser_id',
         'versioningWS' => false,
         'delete' => 'deleted',
         'enablecolumns' => [