+++ CHANGELOG.md
@@ -30,6 +30,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Deprecate `FrontEndUser::getAge` (#1408)
 
 ### Fixed
+- Improve some type annotations (#1453)
 - Validate UIDs in the ORM more strictly (#1402)
 
 ## 5.1.0