+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Add the Crowdin configuration (#1437)
 
 ### Changed
+- Also run the PHP 8.2/8.3 functional tests with the lowest dependencies (#1220)
 
 ### Deprecated